integrations/model-training/pytorch-lightning/pytorch-lightning-hello-world/requirements.txt
comet_ml>=3.44.0
numpy
pytorch-lightning
torch
torchvision
